making client file for dune client
